requirements.txt
attrs==23.2.0
Automat==22.10.0
azure-functions==1.19.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
fake-useragent==1.5.1
filelock==3.14.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
jmespath==1.0.1
lxml==5.2.1
mysql-connector-python==8.4.0
OpenCC==1.1.7
outcome==1.3.0.post0
packaging==24.0
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
PySocks==1.7.1
pywin32==306
queuelib==1.7.0
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapy-selenium==0.0.7
scrapyd==1.4.3
scrapyd-client==1.2.3
selenium==4.20.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tldextract==5.1.2
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.11.0
uberegg==0.1.1
urllib3==2.2.1
w3lib==2.1.2
wsproto==1.2.0
zope.interface==6.3
